Fix error on migration

+++ db/migrate/20210628110058_fix_type_on_decidim_initiatives.rb
@@ -4,19 +4,18 @@ class FixTypeOnDecidimInitiatives < ActiveRecord::Migration[5.2]
   end
 
   def change
-
     # This flag says when mixed and face-to-face voting methods
     # are allowed. If set to false, only online voting will be
     # allowed
     face_to_face_voting_allowed = false
 
-    # rubocop:disable Lint/UnreachableCode
-    add_column :decidim_initiatives_types, :signature_type, :integer, null: false, default: 0
-    # rubocop:enable Lint/UnreachableCode
+    if defined?(InitiativesType) && !(defined?(InitiativesType.signature_type))
+      add_column :decidim_initiatives_types, :signature_type, :integer, null: false, default: 0
+    end
 
     InitiativesType.reset_column_information
 
-    if defined?(Decidim::Initiatives::InitiativesType) && defined?(Decidim::Initiatives::InitiativesType.signature_type)
+    if defined?(InitiativesType) && defined?(InitiativesType.signature_type)
       Decidim::Initiatives::InitiativesType.find_each do |type|
         type.signature_type = if type.online_signature_enabled && face_to_face_voting_allowed
                                 :any
@@ -29,6 +28,8 @@ class FixTypeOnDecidimInitiatives < ActiveRecord::Migration[5.2]
       end
     end
 
-    remove_column :decidim_initiatives_types, :online_signature_enabled
+    if defined?(InitiativesType) && defined?(InitiativesType.online_signature_enabled)
+      remove_column :decidim_initiatives_types, :online_signature_enabled
+    end
   end
 end
